separated ground from rounding corners
